+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## unreleased
+
+### Changed
+
+- Update the default URLs for the secondary endpoints [#253](https://github.com/bugsnag/bugsnag-cli/pull/253)
+
 ## [3.5.0] - 2025-11-12
 
 ### Added